nu met doubles

+++ src/be/vdab/DriekhoekigeTegel.java
@@ -3,17 +3,17 @@ package be.vdab;
 import java.math.BigDecimal;
 
 public class DriekhoekigeTegel extends Tegel{
-    private final int lengte;
-    private final int breedte;
+    private final double lengte;
+    private final double breedte;
 
-    public DriekhoekigeTegel(BigDecimal prijs, int lengte, int breedte) {
+    public DriekhoekigeTegel(BigDecimal prijs, double lengte, double breedte) {
         super(prijs);
         this.lengte = lengte;
         this.breedte = breedte;
     }
 
     @Override
-    public int getOppervlakte() {
+    public double getOppervlakte() {
         return lengte*breedte/2;
     }
 }
